+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#59] Support for extended volume definitions in the `dogu.json`, allowing the creation of kubernetes specific 
+  volumes.
+- [#59] Support for extended service account definitions in the `dogu.json`, allowing the creation of kubernetes 
+  accounts for dogus.
+
+### Removed
+- [#59] Mechanism to patch the generated dogu deployment with custom volumes and service account names. These are now
+  supported by the `dogu.json` and natively generated into the deployment.
 
 ## [v0.15.0] - 2022-11-15
 ### Changed